Batch USB ACM I/O to speed things up
putchar over the USB ACM interface results in USB bulk transfers that
are almost empty: 1 byte used of 64 bytes available. To avoid this
issue, use fwrite with a buffer.

With this, the performance gets quite nice, from 7 minutes down to half
a minute for reading a 16 MiB flash.

 SPI freq. |   time   |  throughput
-----------|----------|--------------
    1 MHz  |  142.5 s |  0.94 Mbit/s
    4 MHz  |   45.7 s |  2.94 Mbit/s
    8 MHz  |   30.5 s |  4.40 Mbit/s
   12 MHz  |   26.3 s |  5.10 Mbit/s
   16 MHz  |   26.1 s |  5.14 Mbit/s
   20 MHz  |   26.0 s |  5.16 Mbit/s

Beyond 12 MHz there are diminishing returns, because the USB "Full Speed"
interface is limited to 12 Mbit/s.